Introduces Webui::Packages::TriggerController
Makes Webui::PackageController smaller and groups those actions
as they clearly belong together.
