+++ src/api/app/views/webui/package/show_actions/_trigger_services.html.haml
@@ -1,5 +1,5 @@
 %li.nav-item
-  = link_to(package_trigger_services_path(package: package, project: project), method: :post, class: 'nav-link',
+  = link_to(services_project_package_trigger_path(project_name: project, package_name: package), method: :post, class: 'nav-link',
             title: 'Trigger Services') do
     %i.fas.fa-project-diagram.fa-lg.me-2
     %span.nav-item-name Trigger Services
